Overhaul theming system and fix Saturday/Sunday attendance weighting
Theming:
- Consolidate four overlapping colour authorities (theme.js, theme-restraint.js,
  theme-chart-bridge.js, charts.js hardcodes) into one: theme.js now owns every
  resolved token, theme-restraint.js is deleted, and apply() dispatches a single
  theme-changed notification instead of double-repainting every chart.
- Replace the flat surface system (surface/surface-muted/surface-soft/card-surface
  all pointing at one raw value) with a real six-step ladder (canvas -> muted ->
  surface -> soft -> card -> raised), and repoint ordinary cards onto the new
  card-surface tier so dialogs/dropdowns read as genuinely more elevated.
- Trim the theme library to 6 neutral light themes plus 5 dark themes inspired by
  well-known editor palettes (Dracula, Gruvbox, Monokai, Everforest, Solarized
  Dark), each with non-blue accents and dark-tuned elevation shadows.
- Fix a chain of contrast bugs found via direct code audit rather than
  screenshot-guessing: a hardcoded `:root { color: #202623 }` fallback that
  theme.js could never override (added `color: var(--ink)` to html/body so
  every unstyled element inherits the real theme ink instead); white text
  assumed readable on any status colour (added computed --*-ink tokens via
  real luminance); chart category-axis labels needlessly muted; KPI labels and
  chart value-axis ticks using the too-dim --muted token instead of --ink-soft;
  a 3px pure-white heatmap cell border reading as a thick grid on dark themes;
  and a dozen leftover hardcoded literal colours/borders from before the theme
  system existed, replaced with theme tokens across app.css, employees.css,
  analytics-extension.css and theme-library.css.
- Fix the "Optimal"/"Underutilized" scatter dots rendering identical colours
  (a positional colour fallback coincidentally matched the same hue as "good"
  in every theme); give Underutilized its own fixed, theme-independent colour.
- Add a WPF DispatcherUnhandledException handler so a future rendering bug
  shows an error dialog instead of silently killing the process (found the
  hard way: an invalid QuickGrid parameter was crashing the app on every visit
  to Timesheets/Imported data).

Attendance and scoring:
- Sunday is now always treated as a weekly off regardless of what the ERP's
  own Duty Type tag says on that row; Saturday counts as a half working day
  (0.5) everywhere day-counts feed scoring (Expected/Filled/Attendance days,
  monthly and weekly), rather than every accountable day counting as a full 1.
  ExpectedTimesheetDays/TimesheetFilledDays (and their weekly equivalents)
  move from int to decimal to carry the half-day weight; new EF migration.
- A week-off row can no longer draw down leave balance even when the ERP also
  tags it Position=Leave (observed on Sundays double-tagged this way) — a day
  nobody could have worked is not a day of leave taken.
- Utilization is now read verbatim from the ERP's own Utilization column
  instead of being recomputed locally as Entered/Compliance, which produced
  different numbers than the ERP under its own tie-breaking rule (prefers
  Billable/Compliance when billable hours exist). Falls back to the old
  formula only for months imported before this column was captured, and only
  when the stored value is 0 with nonzero entered hours (a state the real ERP
  figure can never produce), so historical charts don't show a false 0%.
- New EF migration for the ErpUtilization column; both migrations verified
  against the production DB with a reimport of August 2026 data.

Overview page layout: KPI-row-adjacent chart row moved above the heatmaps,
Attendance & reconciliation narrowed to match Score trend width with
Operational score composition widened accordingly; Engineer performance table
gained more ERP-sourced columns (Code, Utilization, Punch, Variance, Filled)
with colour-coded cells, plus an amber header marker (fixed, theme-independent
colour) distinguishing our own derived scores from values read straight off
the ERP.

+++ src/EngineeringPerformance.UI/Analytics.cs
@@ -300,7 +300,7 @@ public static class Analytics
     /// </summary>
     public sealed record ComplianceRow(
         string Name, string? Code, string? Email, string? TeamName,
-        int ExpectedDays, int FilledDays, int MissingDays,
+        decimal ExpectedDays, decimal FilledDays, decimal MissingDays,
         decimal EnteredHours, decimal ApprovedHours, decimal PendingApprovalHours,
         decimal BillableHours, bool IsNonBillable, bool IsConsultant, bool IsOnProbation)
     {
@@ -322,11 +322,11 @@ public static class Analytics
         return items.Select(item =>
         {
             var employee = employees.FirstOrDefault(e => PersonName.Matches(e.Name, item.EmployeeName));
-            var pending = Math.Max(0, item.EnteredHours - item.ApprovedHours);
+            var pending = Math.Max(0m, item.EnteredHours - item.ApprovedHours);
             return new ComplianceRow(
                 item.EmployeeName, item.EmployeeCode ?? employee?.EmployeeCode, employee?.Email, employee?.TeamName,
                 item.ExpectedTimesheetDays, item.TimesheetFilledDays,
-                Math.Max(0, item.ExpectedTimesheetDays - item.TimesheetFilledDays),
+                Math.Max(0m, item.ExpectedTimesheetDays - item.TimesheetFilledDays),
                 item.EnteredHours, item.ApprovedHours, pending,
                 item.BillableHours,
                 employee?.IsNonBillable ?? false, employee?.IsConsultant ?? false, employee?.IsOnProbation ?? false);
